resolve fix code
