+++ package.json
@@ -1,47 +1,44 @@
 {
-  "name": "examportal",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "@amcharts/amcharts5": "^5.6.2",
-    "@amcharts/amcharts5-fonts": "^5.0.1",
-    "@amcharts/amcharts5-geodata": "^5.1.1",
-    "@auth0/auth0-react": "^2.2.1",
-    "@reduxjs/toolkit": "^1.9.7",
-    "bootstrap": "^5.3.2",
-    "chart.js": "^4.4.1",
-    "file-saver": "^2.0.5",
-    "formik": "^2.4.5",
-    "framer-motion": "^10.16.9",
-    "jsonwebtoken": "^9.0.2",
-    "jws": "^4.0.0",
-    "react": "^18.2.0",
-    "react-bootstrap": "^2.9.1",
-    "react-countdown-now": "^2.1.2",
-    "react-dom": "^18.2.0",
-    "react-icons": "^4.11.0",
-    "react-redux": "^8.1.3",
-    "react-router-dom": "^6.17.0",
-    "react-toastify": "^9.1.3",
-    "safe-buffer": "^5.2.1",
-    "xlsx": "^0.18.5",
-    "yup": "^1.3.2"
-  },
-  "devDependencies": {
-    "@types/react": "^18.2.15",
-    "@types/react-dom": "^18.2.7",
-    "@vitejs/plugin-react-swc": "^3.3.2",
-    "eslint": "^8.45.0",
-    "eslint-plugin-react": "^7.32.2",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.3",
-    "vite": "^4.4.5"
-  }
+    "name": "examportal",
+    "private": true,
+    "version": "0.0.0",
+    "type": "module",
+    "scripts": {
+      "dev": "vite",
+      "build": "vite build",
+      "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
+      "preview": "vite preview"
+    },
+    "dependencies": {
+        "@auth0/auth0-react": "^2.2.1",
+        "@reduxjs/toolkit": "^1.9.7",
+        "bootstrap": "^5.3.2",
+        "chart.js": "^4.4.1",
+        "file-saver": "^2.0.5",
+        "formik": "^2.4.5",
+        "framer-motion": "^10.16.9",
+        "jsonwebtoken": "^9.0.2",
+        "jws": "^4.0.0",
+        "react": "^18.2.0",
+        "react-bootstrap": "^2.9.1",
+        "react-countdown-now": "^2.1.2",
+        "react-dom": "^18.2.0",
+        "react-icons": "^4.11.0",
+        "react-redux": "^8.1.3",
+        "react-router-dom": "^6.17.0",
+        "react-toastify": "^9.1.3",
+        "safe-buffer": "^5.2.1",
+        "xlsx": "^0.18.5",
+        "yup": "^1.3.2"
+      },
+      "devDependencies": {
+        "@types/react": "^18.2.15",
+        "@types/react-dom": "^18.2.7",
+        "@vitejs/plugin-react-swc": "^3.3.2",
+        "eslint": "^8.45.0",
+        "eslint-plugin-react": "^7.32.2",
+        "eslint-plugin-react-hooks": "^4.6.0",
+        "eslint-plugin-react-refresh": "^0.4.3",
+        "vite": "^4.4.5"
+      }
 }
