--- src/pages/home/AdminDashBoardGraph.jsx
@@ -1,18 +1,0 @@
-import React from "react";
-import AdminDashboard from "./admindashboard/AdminDashboard";
-
-
-function AdminDashboardGraph() {
-  
-
-    return (
-      <>
-      <div className=" ">
-     <AdminDashboard />
-     </div>
-      </>
-    )
-  }
-  
-  export default AdminDashboardGraph;
-
